UX-658 Convert TextField to typescript
